Prerel 1.1.0

See merge request personal/wedding-share!7

+++ WeddingShare/Controllers/AdminController.cs
@@ -8,6 +8,10 @@ using WeddingShare.Enums;
 using WeddingShare.Helpers;
 using WeddingShare.Views.Admin;
 using WeddingShare.Models;
+using WeddingShare.Helpers.Database;
+using WeddingShare.Models.Database;
+using System.IO;
+using System.IO.Compression;
 
 namespace WeddingShare.Controllers
 {
@@ -16,19 +20,25 @@ namespace WeddingShare.Controllers
     {
         private readonly IWebHostEnvironment _hostingEnvironment;
         private readonly IConfigHelper _config;
+        private readonly IDatabaseHelper _database;
+        private readonly IImageHelper _imageHelper;
         private readonly ILogger _logger;
         private readonly IStringLocalizer<AdminController> _localizer;
 
         private readonly string UploadsDirectory;
+        private readonly string ThumbnailsDirectory;
 
-        public AdminController(IWebHostEnvironment hostingEnvironment, IConfigHelper config, ILogger<AdminController> logger, IStringLocalizer<AdminController> localizer)
+        public AdminController(IWebHostEnvironment hostingEnvironment, IConfigHelper config, IDatabaseHelper database, IImageHelper imageHelper, ILogger<AdminController> logger, IStringLocalizer<AdminController> localizer)
         {
             _hostingEnvironment = hostingEnvironment;
             _config = config;
+            _database = database;
+            _imageHelper = imageHelper;
             _logger = logger;
             _localizer = localizer;
 
             UploadsDirectory = Path.Combine(_hostingEnvironment.WebRootPath, "uploads");
+            ThumbnailsDirectory = Path.Combine(_hostingEnvironment.WebRootPath, "thumbnails");
         }
 
         [AllowAnonymous]
@@ -38,7 +48,7 @@ namespace WeddingShare.Controllers
             try
             {
                 var passkey = _config.Get("Settings", "Admin_Password");
-                if (string.IsNullOrEmpty(passkey) || string.Equals(passkey, model?.Password))
+                if (string.IsNullOrWhiteSpace(passkey) || string.Equals(passkey, model?.Password))
                 {
                     var claims = new List<Claim>
                     {
@@ -57,7 +67,7 @@ namespace WeddingShare.Controllers
                 _logger.LogError(ex, $"{_localizer["Login_Failed"].Value} - {ex?.Message}");
             }
 
-            return Json(new { success = false, message = "Invalid password" });
+            return Json(new { success = false });
         }
 
         [Authorize]
@@ -69,7 +79,7 @@ namespace WeddingShare.Controllers
         }
 
         [HttpGet]
-        public IActionResult Index()
+        public async Task<IActionResult> Index()
         {
             if (User?.Identity == null || !User.Identity.IsAuthenticated)
             { 
@@ -80,10 +90,19 @@ namespace WeddingShare.Controllers
 
             try
             {
-                if (Directory.Exists(UploadsDirectory))
+                if (!_config.GetOrDefault("Settings", "Single_Gallery_Mode", false))
                 {
-                    model.Galleries = Directory.GetDirectories(UploadsDirectory)?.Select(x => new KeyValuePair<string, string>(Path.GetFileName(x), x))?.ToList();
-                    model.PendingRequests = model.Galleries?.SelectMany(x => Directory.Exists(Path.Combine(x.Value, "Pending")) ? Directory.GetFiles(Path.Combine(x.Value, "Pending"), "*.*", SearchOption.TopDirectoryOnly) : new string[0])?.Select(x => x.Replace(UploadsDirectory, string.Empty))?.ToList();
+                    model.Galleries = await _database.GetAllGalleries();
+                    model.PendingRequests = await _database.GetPendingGalleryItems();
+                }
+                else
+                {
+                    var gallery = await _database.GetGallery("default");
+                    if (gallery != null)
+                    { 
+                        model.Galleries = new List<GalleryModel>() { gallery };
+                        model.PendingRequests = await _database.GetPendingGalleryItems(gallery.Id);
+                    }
                 }
             }
             catch (Exception ex)
@@ -95,23 +114,42 @@ namespace WeddingShare.Controllers
         }
 
         [HttpPost]
-        public IActionResult ReviewPhoto(string galleryId, string photoId, ReviewAction action)
+        public async Task<IActionResult> ReviewPhoto(int id, ReviewAction action)
         {
             if (User?.Identity != null && User.Identity.IsAuthenticated)
             {
                 try
                 {
-                    var galleryDir = Path.Combine(UploadsDirectory, galleryId);
-                    var reviewFile = Path.Combine(galleryDir, "Pending", photoId);
-                    if (System.IO.File.Exists(reviewFile))
-                    {
+                    var review = await _database.GetPendingGalleryItem(id);
+                    if (review != null)
+                    { 
+                        var galleryDir = Path.Combine(UploadsDirectory, review.GalleryName);
+                        var reviewFile = Path.Combine(galleryDir, "Pending", review.Title);
                         if (action == ReviewAction.APPROVED)
                         {
-                            System.IO.File.Move(reviewFile, Path.Combine(galleryDir, Path.GetFileName(reviewFile)));
+                            if (!Directory.Exists(ThumbnailsDirectory))
+                            {
+                                Directory.CreateDirectory(ThumbnailsDirectory);
+                            }
+
+                            await _imageHelper.GenerateThumbnail(reviewFile, Path.Combine(ThumbnailsDirectory, $"{Path.GetFileNameWithoutExtension(reviewFile)}.webp"), _config.GetOrDefault("Settings", "Thumbnail_Size", 720));
+
+                            if (System.IO.File.Exists(reviewFile))
+                            {
+                                System.IO.File.Move(reviewFile, Path.Combine(galleryDir, review.Title));
+                            }
+
+                            review.State = GalleryItemState.Approved;
+                            await _database.EditGalleryItem(review);
                         }
                         else if (action == ReviewAction.REJECTED)
                         {
-                            System.IO.File.Delete(reviewFile);
+                            if (System.IO.File.Exists(reviewFile))
+                            {
+                                System.IO.File.Delete(reviewFile);
+                            }
+
+                            await _database.DeleteGalleryItem(review);
                         }
                         else if (action == ReviewAction.UNKNOWN)
                         {
@@ -128,6 +166,157 @@ namespace WeddingShare.Controllers
                 catch (Exception ex)
                 {
                     _logger.LogError(ex, $"{_localizer["Failed_Reviewing_Photo"].Value} - {ex?.Message}");
+                }
+            }
+
+            return Json(new { success = false });
+        }
+
+        [HttpPost]
+        public async Task<IActionResult> AddGallery(GalleryModel model)
+        {
+            if (User?.Identity != null && User.Identity.IsAuthenticated)
+            {
+                if (!string.IsNullOrWhiteSpace(model?.Name))
+                {
+                    try
+                    {
+                        var check = await _database.GetGallery(model.Name);
+                        if (check == null)
+                        {
+                            return Json(new { success = string.Equals(model?.Name, (await _database.AddGallery(model))?.Name, StringComparison.OrdinalIgnoreCase) });
+                        }
+                        else
+                        { 
+                            return Json(new { success = false, message = _localizer["Gallery_Name_Already_Exists"].Value });
+                        }
+                    }
+                    catch (Exception ex)
+                    {
+                        _logger.LogError(ex, $"{_localizer["Failed_Add_Gallery"].Value} - {ex?.Message}");
+                    }
+                }
+                else
+                { 
+                    return Json(new { success = false, message = _localizer["Name_Cannot_Be_Blank"].Value });
+                }
+            }
+
+            return Json(new { success = false });
+        }
+
+        [HttpPut]
+        public async Task<IActionResult> EditGallery(GalleryModel model)
+        {
+            if (User?.Identity != null && User.Identity.IsAuthenticated)
+            {
+                if (!string.IsNullOrWhiteSpace(model?.Name))
+                {
+                    try
+                    {
+                        var check = await _database.GetGallery(model.Name);
+                        if (check == null || model.Id == check.Id)
+                        {
+                            var gallery = await _database.GetGallery(model.Id);
+                            if (gallery != null)
+                            {
+                                gallery.Name = model.Name;
+                                gallery.SecretKey = model.SecretKey;
+
+                                return Json(new { success = string.Equals(model?.Name, (await _database.EditGallery(gallery))?.Name, StringComparison.OrdinalIgnoreCase) });
+                            }
+                            else
+                            {
+                                return Json(new { success = false, message = _localizer["Failed_Edit_Gallery"].Value });
+                            }
+                        }
+                        else
+                        {
+                            return Json(new { success = false, message = _localizer["Gallery_Name_Already_Exists"].Value });
+                        }
+                    }
+                    catch (Exception ex)
+                    {
+                        _logger.LogError(ex, $"{_localizer["Failed_Edit_Gallery"].Value} - {ex?.Message}");
+                    }
+                }
+                else
+                {
+                    return Json(new { success = false, message = _localizer["Name_Cannot_Be_Blank"].Value });
+                }
+            }
+
+            return Json(new { success = false });
+        }
+
+        [HttpDelete]
+        public async Task<IActionResult> DeleteGallery(int id)
+        {
+            if (User?.Identity != null && User.Identity.IsAuthenticated)
+            {
+                try
+                {
+                    var gallery = await _database.GetGallery(id);
+                    if (gallery != null)
+                    {
+                        var galleryDir = Path.Combine(UploadsDirectory, gallery.Name);
+                        if (Directory.Exists(galleryDir))
+                        {
+                            Directory.Delete(galleryDir, true);
+                        }
+
+                        return Json(new { success = await _database.DeleteGallery(gallery) });
+                    }
+                    else
+                    {
+                        return Json(new { success = false, message = _localizer["Failed_Delete_Gallery"].Value });
+                    }
+                }
+                catch (Exception ex)
+                {
+                    _logger.LogError(ex, $"{_localizer["Failed_Delete_Gallery"].Value} - {ex?.Message}");
+                }
+            }
+
+            return Json(new { success = false });
+        }
+
+        [HttpPost]
+        public async Task<IActionResult> DownloadGallery(int id)
+        {
+            if (User?.Identity != null && User.Identity.IsAuthenticated)
+            {
+                try
+                {
+                    var gallery = await _database.GetGallery(id);
+                    if (gallery != null)
+                    {
+                        var galleryDir = Path.Combine(UploadsDirectory, gallery.Name);
+                        if (Directory.Exists(galleryDir))
+                        {
+                            var tempZipDir = $"Temp";
+                            if (!Directory.Exists(tempZipDir))
+                            { 
+                                Directory.CreateDirectory(tempZipDir);
+                            }
+
+                            var tempZipFile = Path.Combine(tempZipDir, $"{gallery.Name}-{DateTime.Now.ToString("yyyy-MM-dd_HH-mm-ss")}.zip");
+                            ZipFile.CreateFromDirectory(galleryDir, tempZipFile, CompressionLevel.Optimal, false);
+
+                            byte[] bytes = System.IO.File.ReadAllBytes(tempZipFile);
+                            System.IO.File.Delete(tempZipFile);
+
+                            return Json(new { success = true, filename = Path.GetFileName(tempZipFile), content = Convert.ToBase64String(bytes, 0, bytes.Length) });
+                        }
+                    }
+                    else
+                    {
+                        return Json(new { success = false, message = _localizer["Failed_Download_Gallery"].Value });
+                    }
+                }
+                catch (Exception ex)
+                {
+                    _logger.LogError(ex, $"{_localizer["Failed_Download_Gallery"].Value} - {ex?.Message}");
                 }
             }
